+++ src/BoclipsPlayer/BoclipsPlayer.integration.test.ts
@@ -1,3 +1,4 @@
+import { BoclipsAPIError } from './../ErrorHandler/BoclipsPlayerError';
 import MockFetchVerify from '../test-support/MockFetchVerify';
 import { VideoResourceFactory } from '../test-support/TestFactories';
 import { BoclipsPlayer } from './BoclipsPlayer';
@@ -5,15 +6,14 @@ import { BoclipsPlayer } from './BoclipsPlayer';
 jest.unmock('plyr');
 
 let container: HTMLElement;
-let player: BoclipsPlayer;
 
 beforeEach(() => {
   container = document.createElement('div');
   document.body.appendChild(container);
-  player = new BoclipsPlayer(container);
 });
 
 it('remove an error when reloading a good video', async () => {
+  const player = new BoclipsPlayer(container);
   const errorUri = 'http://server/path/to/error/video';
   MockFetchVerify.get(errorUri, {}, 404);
 
@@ -26,4 +26,23 @@ it('remove an error when reloading a good video', async () => {
   // User "oh no!"
   await player.loadVideo(goodUri);
   expect(container.querySelector('.error')).toBeFalsy();
+});
+
+it('calls a passed in error handler when something goes wrong', async () => {
+  const onErrorCallback = jest.fn();
+  const player = new BoclipsPlayer(container);
+  player.onError(onErrorCallback);
+
+  const errorUri = 'http://server/path/to/error/video';
+  MockFetchVerify.get(errorUri, {}, 404);
+
+  await player.loadVideo(errorUri);
+
+  const expectedError: BoclipsAPIError = {
+    type: 'API_ERROR',
+    message: 'Error retrieving video from API',
+    statusCode: 404,
+  };
+
+  expect(onErrorCallback).toHaveBeenCalledWith(expectedError);
 });