+++ src/ErrorHandler/__mocks__/ErrorHandler.ts
@@ -6,5 +6,6 @@ export const ErrorHandler = jest.fn().mockImplementation(() => {
     clearError: jest.fn(),
     handleError: jest.fn(),
     isDefinedError: jest.fn().mockReturnValue(false),
+    onError: jest.fn(),
   } as ErrorHandlerInstance;
 });