+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.8.0] - 2021-08-11
+
+- Add an onError call back to the player
+```
+const player = PlayerFactory.get(playerContainer);
+player.onError((error: BoclipsError) => {
+  // do something with error
+  
+  // remove the error element from the player
+  player.getErrorHandler().clearError()
+  
+  // to show a new one error message, you'll have to manipulate the DOM
+});
+```
+
 ## [6.7.2] - 2021-08-02
 
 - Add tooltips to seek and buttons